// backend/src/dtos/request/pet/update-pet.dto.ts
export interface UpdatePetDTO {
  // Nombre de la mascota (opcional)
  name?: string;
  // Especie de la mascota (opcional)
  species?: string;
  // Raza de la mascota (opcional)
  breed?: string;
  // Edad de la mascota en años (opcional)
  age?: number;
  // Nombre del dueño (opcional)
  ownerName?: string;
}